src/main/java/net/amygdalum/comtemplate/engine/expressions/EvalContextVar.java
package net.amygdalum.comtemplate.engine.expressions;

import net.amygdalum.comtemplate.engine.Scope;
import net.amygdalum.comtemplate.engine.TemplateExpression;
import net.amygdalum.comtemplate.engine.TemplateExpressionVisitor;

public class EvalContextVar implements TemplateExpression {

	private String name;

	public EvalContextVar(String name) {
		this.name = name;
	}
	
	public String getName() {
		return name;
	}
	
	@Override
	public <T> T apply(TemplateExpressionVisitor<T> visitor, Scope scope) {
		return visitor.visitEvalContextVar(this, scope);
	}

	@Override
	public String toString() {
		return "@@" + name;
	}
}
